# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pharmacy POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierListResponse"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{supplier_id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [suppliers]
      summary: List purchase orders
      parameters:
        - name: supplier_id
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderListResponse"
    post:
      operationId: createPurchaseOrder
      tags: [suppliers]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /suppliers/purchase-orders/{po_id}:
    get:
      operationId: getPurchaseOrder
      tags: [suppliers]
      summary: Get purchase order
      parameters:
        - name: po_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /suppliers/grn:
    post:
      operationId: createGrn
      tags: [suppliers]
      summary: Create GRN (Goods Received Note)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrnInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grn"
    get:
      operationId: listGrns
      tags: [suppliers]
      summary: List GRNs
      parameters:
        - name: supplier_id
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: GRNs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grn"
  /suppliers/payables-aging:
    get:
      operationId: getPayablesAging
      tags: [suppliers]
      summary: Get payables aging report
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
        - name: as_of_date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Aging report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayablesAging"

  # ── DASHBOARD ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
      required:
        - page
        - per_page
        - total
        - pages
        - has_next
        - has_prev

    # ── AUTH ──────────────────────────────────────────────────
...
    Supplier:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contact_person:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        license_number:
          type: ["string", "null"]
        credit_days:
          type: ["integer", "null"]
        payment_terms:
          type: ["string", "null"]
        is_active:
          type: boolean
        outstanding_balance:
          type: ["number", "null"]
        created_at:
          type: string
      required:
        - id
        - name
        - is_active
        - created_at
    SupplierInput:
      type: object
      properties:
        name:
          type: string
        contact_person:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        license_number:
          type: ["string", "null"]
        credit_days:
          type: ["integer", "null"]
        payment_terms:
          type: ["string", "null"]
      required:
        - name
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        contact_person:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        credit_days:
          type: ["integer", "null"]
    SupplierListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Supplier"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    PurchaseOrder:
      type: object
      properties:
        id:
          type: string
        po_number:
          type: string
        supplier_id:
          type: string
        supplier_name:
          type: string
        branch_id:
          type: string
        status:
          type: string
        order_date:
          type: string
        expected_date:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            properties:
              medicine_id:
                type: string
              medicine_name:
                type: string
              quantity:
                type: integer
              unit_price:
                type: number
              total:
                type: number
            required:
              - medicine_id
              - medicine_name
              - quantity
              - unit_price
              - total
        total_amount:
          type: number
        notes:
          type: ["string", "null"]
        created_at:
          type: string
      required:
        - id
        - po_number
        - supplier_id
        - supplier_name
        - branch_id
        - status
        - order_date
        - items
        - total_amount
        - created_at
    PurchaseOrderInput:
      type: object
      properties:
        po_number:
          type: ["string", "null"]
        supplier_id:
          type: string
        branch_id:
          type: string
        order_date:
          type: string
        expected_date:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            properties:
              medicine_id:
                type: string
              quantity:
                type: integer
              unit_price:
                type: number
            required:
              - medicine_id
              - quantity
              - unit_price
        notes:
          type: ["string", "null"]
      required:
        - supplier_id
        - branch_id
        - order_date
        - items
    PurchaseOrderListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrder"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    Grn:
      type: object
      properties:
        id:
          type: string
        grn_number:
          type: string
        po_id:
          type: ["string", "null"]
        supplier_id:
          type: string
        supplier_name:
          type: string
        branch_id:
          type: string
        received_date:
          type: string
        supplier_invoice_number:
          type: ["string", "null"]
        total_amount:
          type: number
        created_at:
          type: string
      required:
        - id
        - grn_number
        - supplier_id
        - supplier_name
        - branch_id
        - received_date
        - total_amount
        - created_at
    GrnInput:
      type: object
      properties:
        grn_number:
          type: ["string", "null"]
        po_id:
          type: ["string", "null"]
        supplier_id:
          type: string
        branch_id:
          type: string
        received_date:
          type: string
        supplier_invoice_number:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            properties:
              medicine_id:
                type: string
              batch_number:
                type: string
              expiry_date:
                type: string
              quantity_received:
                type: integer
              purchase_price:
                type: number
              selling_price:
                type: number
            required:
              - medicine_id
              - batch_number
              - expiry_date
              - quantity_received
              - purchase_price
              - selling_price
        notes:
          type: ["string", "null"]
      required:
        - supplier_id
        - branch_id
        - received_date
        - items
    PayablesAging:
      type: object
      properties:
        supplier_id:
          type: string
        supplier_name:
          type: string
        current:
          type: number
        days_0_30:
          type: number
        days_31_60:
          type: number
        days_61_90:
          type: number
        days_90_plus:
          type: number
        total_outstanding:
          type: number
      required:
        - supplier_id
        - supplier_name
        - current
        - days_0_30
        - days_31_60
        - days_61_90
        - days_90_plus
        - total_outstanding

    # ── DASHBOARD ──────────────────────────────────────────────────
...